Rename 'client_exporter' to 'exporter'
