+++ js/bundles/modules/core.js
@@ -10,7 +10,7 @@ if(DevExpress._DEVEXTREME_BUNDLE_INITIALIZED) {
 }
 DevExpress._DEVEXTREME_BUNDLE_INITIALIZED = true;
 
-DevExpress.clientExporter = require("../../client_exporter");
+DevExpress.clientExporter = require("../../exporter");
 
 DevExpress.VERSION = require("../../core/version");
