+++ js/ui/data_grid/ui.data_grid.export.js
@@ -6,7 +6,7 @@ var $ = require("../../core/renderer"),
     arrayUtils = require("../../core/utils/array"),
     dataGridCore = require("./ui.data_grid.core"),
     exportMixin = require("../grid_core/ui.grid_core.export_mixin"),
-    clientExporter = require("../../client_exporter"),
+    clientExporter = require("../../exporter"),
     messageLocalization = require("../../localization/message"),
     excelExporter = clientExporter.excel,
     Button = require("../button"),
