+++ js/ui/pivot_grid/ui.pivot_grid.export.js
@@ -4,7 +4,7 @@ var Class = require("../../core/class"),
     each = require("../../core/utils/iterator").each,
     formatHelper = require("../../format_helper"),
     numberLocalization = require("../../localization/number"),
-    clientExporter = require("../../client_exporter"),
+    clientExporter = require("../../exporter"),
     excelExporter = clientExporter.excel,
     DEFAULT_DATA_TYPE = "string",
     exportMixin = require("../grid_core/ui.grid_core.export_mixin"),
